requirements.txt
python-telegram-bot[job-queue]>=21.0,<22.0
openai>=1.30,<2.0
aiosqlite>=0.20,<1.0
python-dotenv>=1.0,<2.0
httpx>=0.27,<1.0
